ios: wire-up error to platform callbacks (#410)
Description: exposing the contents of envoy_error up to the platform. Note that this implementation differs slightly from Android. I have opened #409 to consolidate both platforms.
Risk Level:
Testing: local, and CI

+++ mobile/library/swift/src/EnvoyError.swift
@@ -3,13 +3,13 @@ import Foundation
 @objcMembers
 public final class EnvoyError: NSObject, Error {
   /// Error code associated with the exception that occurred.
-  public let errorCode: Int
+  public let errorCode: UInt64
   /// A description of what exception that occurred.
   public let message: String
   /// Optional cause for the error.
   public let cause: Error?
 
-  init(errorCode: Int, message: String, cause: Swift.Error?) {
+  init(errorCode: UInt64, message: String, cause: Swift.Error?) {
     self.errorCode = errorCode
     self.message = message
     self.cause = cause
